fix HW8 after review

+++ src/main/webapp/resources/js/topjava.common.js
@@ -3,12 +3,6 @@ let context, form;
 function makeEditable(ctx) {
     context = ctx;
     form = $('#detailsForm');
-    $(".delete").click(function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).attr("id"));
-        }
-    });
-
     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
         failNoty(jqXHR);
     });
@@ -23,26 +17,19 @@ function add() {
 }
 
 function deleteRow(id) {
-    $.ajax({
-        url: context.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTable();
-        successNoty("Deleted");
-    });
+    if (confirm('Are you sure?')) {
+        $.ajax({
+            url: context.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            context.updateTable();
+            successNoty("Deleted");
+        });
+    }
 }
 
-function updateTable() {
-    let sd = $("input[name='startDate']").val() ? $("input[name='startDate']").val() : "";
-    let ed = $("input[name='endDate']").val() ? $("input[name='endDate']").val() : "";
-    let st = $("input[name='startTime']").val() ? $("input[name='startTime']").val() : "";
-    let et = $("input[name='endTime']").val() ? $("input[name='endTime']").val() : "";
-
-
-    $.get(context.ajaxUrl + "filter?startDate=" + sd + "&endDate=" + ed
-        + "&startTime=" + st + "&endTime=" + et, function (data) {
-        context.datatableApi.clear().rows.add(data).draw();
-    });
+function updateTableByData(data) {
+    context.datatableApi.clear().rows.add(data).draw();
 }
 
 function save() {
@@ -52,7 +39,7 @@ function save() {
         data: form.serialize()
     }).done(function () {
         $("#editRow").modal("hide");
-        updateTable();
+        context.updateTable();
         successNoty("Saved");
     });
 }
